SyncAudit/requirements.txt
# SyncAudit - Universal Sync Audit Service
# Data flow visibility for any project → external system integration

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Dashboard
streamlit==1.31.0

# HTTP Client (for external API verification)
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Security
python-multipart==0.0.6

# Export formats
pandas==2.2.0

# Dev/Testing
pytest==8.0.0
pytest-asyncio==0.23.4

# Production (Railway)
gunicorn==21.2.0
